Follow PEP 8 - Style Guide for Python Code
PEP 8 is a style guide which is widely agreed and used in the Python
community. Follow it could encourage more Python developers to join
the development.

This commit changes the code to follow all rules of PEP 8, except
E501 - the 80 chars limitation of the each lines. For modern programs,
120 chars is more comfortable.

There are 181 violations of PEP 8 in the project:

$ flake8 setup.py weibo.py snspy.py | grep -v E501 | wc -l
181

All of the violations have been fixed:

$ flake8 setup.py weibo.py snspy.py | grep -v E501 | wc -l
0

Signed-off-by: Tom Li <biergaizi@members.fsf.org>